21f1005675assignment_week7/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-pubsub
google-api-core
google-cloud-functions
google-auth